backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.0  # Modern psycopg3 with binary wheels
neo4j>=5.16.0
alembic>=1.13.1

# Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Graph and Analytics
networkx>=3.2.1
numpy>=1.26.3
pandas>=2.1.4

# Causal Inference (optional - install if needed)
# dowhy>=0.11.1
# pgmpy>=0.1.24
# econml>=0.14.1
scikit-learn>=1.4.0

# PDF Generation
reportlab>=4.0.8
# weasyprint>=60.2  # Requires system dependencies

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
